backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
opencv-python==4.9.0.80
numpy==1.26.4
torch==2.2.1
ultralytics==8.1.27
paho-mqtt==1.6.1
python-dotenv==1.0.1 